requirements.txt
pillow
tqdm
dataclasses